fix: only show login with invalid session
There was a problem with the user always seeing the login screen every
time they open the app. This is probably because there was no
`isLoading` logic for session. There is currently no expiration date for
authentication tokens according to Supabase, so this is probably the
issue.

Current authentication settings:
Time-box user sessions: 0
Inactivity timeout: 0